메뉴_crud_기능_추가 : feat : {가격 BigDecimal -> Integer ,카테고리 id Integer -> Long , 카멜표기법, url 경로 케밥 케이스, jpa auditing, log 처리 } https://github.com/CampusTable/campus-table-be/issues/7

+++ src/main/java/com/campustable/be/domain/menu/dto/MenuRequest.java
@@ -10,9 +10,9 @@ import java.math.BigDecimal;
 @NoArgsConstructor
 public class MenuRequest {
 
-    private Integer categoryId;
+    private Long categoryId;
     private String menuName;
-    private BigDecimal price;
+    private Integer price;
     private String menuPicture;
     private Boolean available;
     private Integer stockQuantity;
