+++ src/main/java/com/campustable/be/domain/menu/repository/MenuRepository.java
@@ -2,7 +2,6 @@ package com.campustable.be.domain.menu.repository;
 
 import com.campustable.be.domain.menu.entity.Menu;
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
 
 import java.util.List;
 import java.util.Optional;
@@ -10,8 +9,8 @@ import java.util.Optional;
 
 public interface MenuRepository extends JpaRepository<Menu,Long>{
 
-    Optional<Menu> findByCategoryIdAndMenuName(Integer categoryId, String menuName);
+    Optional<Menu> findByCategoryIdAndMenuName(Long categoryId, String menuName);
 
-    List<Menu> findByCategoryId(Integer categoryId);
+    List<Menu> findByCategoryId(Long categoryId);
 
 }
